Add properties for AlignExprs, make content->text
Items and Metadata could specify "content", which is not the content AlignmentExpression. This has been changed to "text", as "content" is now the value of the content alignment expression attribute (similarly, "alignment" and "segmentation" are now accessible from Items and Tiers as properties, though they may to the internal attributes dictionary).

I also separated the code for attribute defaults and inheritance to its own XigtAttributeMixin, which also handles the setting of id and type values.