Add top-level import and export scripts, latex out
This commit improves the latex formatter, though it is still pretty limited.

The top-level scripts import.py and export.py act as interfaces to the modules in importers/ and exporters/. Not only is it more convenient, but avoids import problems (such as the Xigt module 'toolbox' shadowing the 'toolbox' package elsewhere in the system).