Add query.ancestors() and query.descendants(). Resolves #15.
Both ancestors() and descendants() take an object (either a tier
or an item) and recursively find all other objects that it refers
to or that refer to it, respectively. They yield tuples:
  (sourcetier, refattr, reftier, refitems)

This commit also adds ref.dereference() and ref.dereference_all()
methods (same name as a previous, non-official function, but
slightly different semantics). These take an object (tier or item)
and a reference attribute, and return the object(s) referred by
that reference attribute.