Add support for namespaces and <meta> children.
This commit significantly reworks the XML serializer. A lot of work is still being manually done, particularly with regards to namespaces, but this change relies on the ElementTree module more than before for serialization.

The MetaChild class in xigt.metadata is a basic container for XML that occurs under a <meta> element. In the previous revisions of Xigt, this XML was just reserialized and stored as a string, but then it couldn't be reformatted for different indentation levels and couldn't be queried structurally. Now it can more easily be examined like other Xigt data structures.

Namespaces in Xigt are also very XML-centric, but now they can at least be transformed to apply to other output formats, if necessary. Anything inheriting from XigtAttributeMixin (which is all classes) now has a 'namespace' attribute and a 'nsmap' attribute. The 'namespace' attribute is the URI namespace of the current element. The 'nsmap' attribute is a dictionary mapping prefixes to URIs, and is used during serialization to recover the nice prefixes instead of 'ns0', 'ns1', etc.