Robustify xigtxml codec.
XigtXML, unfortunately, no longer uses the incremental XML writer because of difficulties with compatibility and generating correct XML. It may be re-enabled in the future. This shouldn't be a problem with reasonably sized XML (I tested up to 16Mb), but there may be memory issues with running the huge files (200Mb+). However, the upcoming release of the ODIN corpus may not even include the huge file, so this might not come up. In any case, I do want it re-enabled in the future.