Fix ODIN import regarding metadata. Resolves #29.
This commit also changes namespace behavior so if a nsmap on an object is None, it looks to its parent's nsmap. If the nsmap is a dict (even {}) it uses its own.

Also, IGTs yielded in 'transient' mode will now have their _parent attribute set so inheritance of attributes, metadata, and nsmap will work.