Make XigtPath work with namespaces.
Namespaces on things like MetaChild objects are now appropriately mapped using the object's nsmap. This way, queries like igt/metadata//dc:subject/@olac:code will work as expected.

This commit also changes the behavior of XigtContainerMixin.select() so that any model attribute (like id, type, etc.) can be used as criteria for selection. Namespaces are replaced with the URL if a mapping is found on nsmap.