Fix #35: replace `xigt process` with `partition`
The `xigt process` command is now deprecated (or reserved for future use, perhaps). The `xigt partition` command replaces its functionality for splitting a corpus by some value, and this value can now be selected with a XigtPath.

The `xigt sort` command is also added for sorting Igts, Tiers, and Items by XigtPath keys.