Improve performance and fix XigtXml bug.
The calcuation of allowed_reference_attributes() and creation of indices was severly slowing down processing of large corpora, so the've been optimized to avoid unnecessary function calls and loops.

And a bug in XigtXml where Igts were not being processed with nsiterparse() (and thus didn't have namespace support) was causing xigtxml.loads() to fail. The loads() function now calls the load() function, so this bug is fixed, but the underlying bug (in default_decode_xigtcorpus() may persist.