Add --count option to `xigt sort` command
Non-hashable objects couldn't be counted with the `--unique` option, so the `count` option doesn't hash anything; it just counts matches. The `--unique` option worked before with Igt objects because they didn't define an __eq__() method, so hashing just operated on the Python object ID, but now __eq__() is defined so it no longer works for these objects.

I added the `--count` option instead of defining a __hash__() method because the fact that they were hashable before was an accident.